+++ test/WebSites/FilesWebSite/Controllers/EmbeddedFilesController.cs
@@ -11,7 +11,7 @@ namespace FilesWebSite
     {
         public IActionResult DownloadFileWithFileName()
         {
-            return new FilePathResult("/Greetings.txt", "text/plain")
+            return new VirtualFilePathResult("/Greetings.txt", "text/plain")
             {
                 FileProvider = new EmbeddedFileProvider(GetType().GetTypeInfo().Assembly, "FilesWebSite.EmbeddedResources"),
                 FileDownloadName = "downloadName.txt"
